update all tests
