+++ packages/basis/src/index.ts
@@ -1,6 +1,4 @@
 export * from './Basis';
-export * from './loader/BasisParser';
-export * from './loader/detectBasis';
-export * from './loader/loadBasis';
+export * from './loader';
 export * from './TranscoderWorker';
