+++ packages/basis/test/basis.tests.ts
@@ -1,19 +1,22 @@
-// CRASHES JEST - ELECTRON
-// it.only('should load a a Basis image', async () =>
-// {
-//     await BasisLoader.loadTranscoder(
-//         'https://cdn.jsdelivr.net/npm/@pixi/basis@6.3.2/assets/basis_transcoder.js',
-//         'https://cdn.jsdelivr.net/npm/@pixi/basis@6.3.2/assets/basis_transcoder.wasm'
-//     );
+describe('Basis loading', () =>
+{
+    // CRASHES JEST - ELECTRON
+    it.skip('should load a a Basis image', async () =>
+    {
+        // await BasisLoader.loadTranscoder(
+        //     'https://cdn.jsdelivr.net/npm/@pixi/basis@6.3.2/assets/basis_transcoder.js',
+        //     'https://cdn.jsdelivr.net/npm/@pixi/basis@6.3.2/assets/basis_transcoder.wasm'
+        // );
 
-//     //  console.log('DOING!!');
-//     const loader = new Loader();
+        // //  console.log('DOING!!');
+        // const loader = new Loader();
 
-//     loader['_parsers'].push(loadBasis);
+        // loader['_parsers'].push(loadBasis);
 
-//     const texture: Texture = await loader.load(`https://pixijs.io/compressed-textures-example/images/kodim20.basis`);
+        // const texture: Texture = await loader.load(`https://pixijs.io/compressed-textures-example/images/kodim20.basis`);
 
-//     // expect(texture.baseTexture.valid).toBe(true);
-//     // expect(texture.width).toBe(1000);
-//     // expect(texture.height).toBe(664);
-// });
+        // expect(texture.baseTexture.valid).toBe(true);
+        // expect(texture.width).toBe(1000);
+        // expect(texture.height).toBe(664);
+    });
+});
